+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 + Changed user page URLs (contributions page) from `/u/user` to `/@user`.
 + `/manage/` is now a list of all managing options, including export/import and the brand new 
   `/manage/accounts`.
++ Users can now be disabled (and re-enabled) by administrator.
 + TOC is now shown in pages when screen width is greater than 960 pixels.
 + Style changes: added a top bar with the site title. It replaces the floating menu on the top right.
 + Now logged-in users have an “Edit” button below the first heading. All users can access page history